refactor(api): split backend/src/api/training.rs into multiple modules
The massive backend/src/api/training.rs file has been split into multiple smaller modules (exercises, energy, measurements, muscles, plans, plates, sessions, sets, stats, and shared types) inside backend/src/api/training/ to improve readability and maintainability.
No functional changes were made. All tests pass.
