+++ backend/src/api/training/mod.rs
@@ -1,0 +1,21 @@
+pub mod energy;
+pub mod exercises;
+pub mod measurements;
+pub mod muscles;
+pub mod plans;
+pub mod plates;
+pub mod sessions;
+pub mod sets;
+pub mod shared;
+pub mod stats;
+
+pub use energy::*;
+pub use exercises::*;
+pub use measurements::*;
+pub use muscles::*;
+pub use plans::*;
+pub use plates::*;
+pub use sessions::*;
+pub use sets::*;
+pub use shared::*;
+pub use stats::*;
